use sinf instead of sin

+++ texture.h
@@ -24,7 +24,7 @@ public:
 	checker_texture() {}
 	checker_texture(texture *t0, texture *t1) : even(t0), odd(t1) {}
 	virtual vec3 value(float u, float v, const vec3 &p) const {
-		float sines = sinf(10.0f * p.x()) * sinf(10.0f * p.y()) * sin(10.0f * p.z());
+		float sines = sinf(10.0f * p.x()) * sinf(10.0f * p.y()) * sinf(10.0f * p.z());
 		if (sines < 0.0f) {
 			return odd->value(u, v, p);
 		} else {
